Fixed Warnings Working Login / HomePage with set credentials
