+++ Key-Commands.md
@@ -1,3 +1,7 @@
 To Start the Project
 
 npm start
+
+To Start the Project without Cache and the latest changes
+
+npx expo start -c
