+++ app/LoginPage/styles.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: "center",
+        padding: 16,
+    },
+    title: {
+        fontSize: 24,
+        fontWeight: "bold",
+        marginBottom: 16,
+        textAlign: "center",
+    },
+    input: {
+        borderWidth: 1,
+        borderColor: "#ccc",
+        padding: 8,
+        marginBottom: 16,
+        borderRadius: 4,
+    },
+});
+
+export default styles;
