+++ app/index.tsx
@@ -1,4 +1,16 @@
-import { registerRootComponent } from "expo";
-import App from "./App";
+import React from "react";
+import { Redirect } from "expo-router";
+import useAuthStore from "./stores/authStore";
 
-registerRootComponent(App);
+const IndexPage = () => {
+    const { authenticated } = useAuthStore();
+
+    // Adjust paths to match your current structure
+    return authenticated ? (
+        <Redirect href="/HomePage" />
+    ) : (
+        <Redirect href="/LoginPage" />
+    );
+};
+
+export default IndexPage;
